Align Rust SDK with openapi.sdk.json spec
Sync models and API functions to match the current OpenAPI specification:
- Remove orphaned endpoints (health, api_keys) and models not in spec
- Add visibility support (field on images, upload param, update endpoint)
- Add ActiveSignedUrl, Credits, UpdateVisibility request/response models
- Fix integer types (i32 → i64) for sizes, timestamps, and plan limits
- Make nullable fields Optional (plan_limits, subscription_ends_at, preset description)
- Fix delete_preset parameter from id to name per spec
- Add missing error variants (Status403, Status404, Status429)

+++ src/models/current_period.rs
@@ -18,16 +18,16 @@ pub struct CurrentPeriod {
     pub period: String,
     /// Unix timestamp of period start
     #[serde(rename = "period_start")]
-    pub period_start: i32,
+    pub period_start: i64,
     /// Unix timestamp of period end
     #[serde(rename = "period_end")]
-    pub period_end: i32,
+    pub period_end: i64,
     /// Uploads this period
     #[serde(rename = "uploads")]
     pub uploads: i32,
     /// Bandwidth used this period in bytes
     #[serde(rename = "bandwidth_bytes")]
-    pub bandwidth_bytes: i32,
+    pub bandwidth_bytes: i64,
     /// API requests this period
     #[serde(rename = "api_requests")]
     pub api_requests: i32,
@@ -39,10 +39,10 @@ pub struct CurrentPeriod {
 impl CurrentPeriod {
     pub fn new(
         period: String,
-        period_start: i32,
-        period_end: i32,
+        period_start: i64,
+        period_end: i64,
         uploads: i32,
-        bandwidth_bytes: i32,
+        bandwidth_bytes: i64,
         api_requests: i32,
         transformations: i32,
     ) -> CurrentPeriod {
